+++ testing/test_full.py
@@ -30,6 +30,7 @@ from GANDLF.cli import (
     run_deployment,
     recover_config,
     post_training_model_optimization,
+    generate_metrics_dict,
 )
 from GANDLF.schedulers import global_schedulers_dict
 from GANDLF.optimizers import global_optimizer_dict
@@ -1138,6 +1139,12 @@ def test_train_metrics_segmentation_rad_2d(device):
         "hausdorff",
         "hausdorff95",
         "normalized_surface_dice",
+        "sensitivity",
+        "sensitivity_per_label",
+        "specificity_segmentation",
+        "specificity_segmentation_per_label",
+        "jaccard",
+        "jaccard_per_label",
     ]
     parameters["model"]["architecture"] = "resunet"
     parameters["model"]["onnx_export"] = False
@@ -1891,6 +1898,12 @@ def test_train_checkpointing_segmentation_rad_2d(device):
         "hd100_per_label",
         "normalized_surface_dice",
         "normalized_surface_dice_per_label",
+        "sensitivity",
+        "sensitivity_per_label",
+        "specificity_segmentation",
+        "specificity_segmentation_per_label",
+        "jaccard",
+        "jaccard_per_label",
     ]
     parameters["model"]["architecture"] = "unet"
     parameters["model"]["onnx_export"] = False
@@ -1939,11 +1952,7 @@ def test_generic_model_patch_divisibility():
     parameters["model"]["amp"] = True
     parameters["model"]["print_summary"] = False
     parameters["model"]["num_channels"] = 3
-    parameters["metrics"] = [
-        "dice",
-        "hausdorff",
-        "hausdorff95" "normalized_surface_dice_per_label",
-    ]
+    parameters["metrics"] = ["dice"]
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
 
     # this assertion should fail
@@ -2997,3 +3006,49 @@ def test_generic_random_numbers_are_deterministic_on_cpu():
     assert f == h
 
     print("passed")
+
+
+def test_generic_cli_function_metrics_cli_rad_nd():
+    print("49: Starting metric calculation tests")
+    for dim in ["2d", "3d"]:
+        for problem_type in ["segmentation", "classification"]:
+            # read and parse csv
+            training_data, _ = parseTrainingCSV(
+                inputDir + f"/train_{dim}_rad_{problem_type}.csv"
+            )
+            if problem_type == "segmentation":
+                labels_array = training_data["Label"]
+            else:
+                labels_array = training_data["ValueToPredict"]
+            training_data["target"] = labels_array
+            training_data["prediction"] = labels_array
+
+            # read and initialize parameters for specific data dimension
+            parameters = parseConfig(
+                testingDir + f"/config_{problem_type}.yaml", version_check_flag=False
+            )
+            parameters["modality"] = "rad"
+            parameters["patch_size"] = patch_size["2D"]
+            parameters["model"]["dimension"] = 2
+            if dim == "3d":
+                parameters["patch_size"] = patch_size["3D"]
+                parameters["model"]["dimension"] = 3
+
+            parameters["verbose"] = False
+
+            temp_infer_csv = os.path.join(outputDir, "temp_csv.csv")
+            training_data.to_csv(temp_infer_csv, index=False)
+
+            output_file = os.path.join(outputDir, "output.yaml")
+            training_data.to_csv(temp_infer_csv, index=False)
+
+            temp_config = get_temp_config_path()
+            with open(temp_config, "w") as file:
+                yaml.dump(parameters, file)
+
+            # run the metrics calculation
+            generate_metrics_dict(temp_infer_csv, temp_config, output_file)
+
+            assert os.path.isfile(output_file), "Metrics output file was not generated"
+
+            sanitize_outputDir()